+++ Formula/accio.rb
@@ -1,7 +1,7 @@
 class Accio < Formula
   desc "Dependency manager driven by SwiftPM for iOS & Co."
   homepage "https://github.com/JamitLabs/Accio"
-  url "https://github.com/JamitLabs/Accio.git", :tag => "0.2.1", :revision => "bbe25288b7fa61dc2a70acee70285266b8abd804"
+  url "https://github.com/JamitLabs/Accio.git", :tag => "0.2.2", :revision => "2ac691f722cd36c34f92bc1d59337a3fa28a43c0"
   head "https://github.com/JamitLabs/Accio.git"
 
   depends_on :xcode => ["10.0", :build]