+++ Sources/AccioKit/Models/Manifest.swift
@@ -8,7 +8,7 @@ class Manifest: Decodable {
     // MARK: - Sub Types
     struct Target: Decodable {
         struct Dependency: Decodable {
-            let name: String
+            let byName: [String]
         }
 
         let name: String
@@ -17,8 +17,12 @@ class Manifest: Decodable {
     }
 
     struct Product: Decodable {
+        struct ProductType: Decodable {
+            let library: [String]?
+        }
+
         let name: String
-        let productType: String
+        let type: ProductType
         let targets: [String]
     }
 
@@ -30,11 +34,11 @@ class Manifest: Decodable {
 
 extension Manifest {
     var appTargets: [AppTarget] {
-        return targets.filter { $0.type == "regular" }.map { AppTarget(projectName: name, targetName: $0.name, dependentLibraryNames: $0.dependencies.map { $0.name }) }
+        return targets.filter { $0.type == "regular" }.map { AppTarget(projectName: name, targetName: $0.name, dependentLibraryNames: $0.dependencies.flatMap { $0.byName }) }
     }
 
     func frameworkDependencies(ofLibrary libraryName: String, dependencyGraph: DependencyGraph) throws -> [Framework] {
-        let libraryProducts: [Product] = products.filter { $0.productType == "library" }
+        let libraryProducts: [Product] = products.filter { $0.type.library != nil && !$0.type.library!.isEmpty }
 
         guard let product: Product = libraryProducts.first(where: { $0.name == libraryName }) else {
             print("Manifest: Could not find library product with name '\(libraryName)' in package manifest for project '\(name)'.", level: .error)
@@ -42,7 +46,7 @@ extension Manifest {
         }
 
         let productsTargets: [Target] = targets.filter { product.targets.contains($0.name) }
-        let dependencyTargetNames: [String] = Array(Set(productsTargets.flatMap { $0.dependencies.map { $0.name } })).sorted()
+        let dependencyTargetNames: [String] = Array(Set(productsTargets.flatMap { $0.dependencies.flatMap { $0.byName } })).sorted()
 
         let projectTargetNames: [String] = targets.map { $0.name }
         let projectProductNames: [String] = products.map { $0.name }