+++ CHANGELOG.md
@@ -3,15 +3,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.3.0]
 ### Added
 - None.
 ### Changed
-- None.
+- Add support for Swift 5 and Xcode 10.2.
+- Separate cached frameworks by Swift tools version.
 ### Deprecated
 - None.
 ### Removed
-- None.
+- Drop support for Swift 4.2 and Xcode <=10.1.
 ### Fixed
 - None.
 ### Security